Update code structure and css
