+++ src/App.js
@@ -1,0 +1,13 @@
+import Calculate from './components/calculator';
+import DisplayQuote from './components/DisplayQuote';
+
+function App() {
+  return (
+    <>
+      <DisplayQuote />
+      <Calculate />
+    </>
+  );
+}
+
+export default App;
