+++ src/components/QuoteData.js
